pyproject.toml
[project]
name = "finops-as-code-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=15.4.0",
    "fastapi[standard]>=0.119.0",
    "google-auth>=2.41.1",
    "pydantic-settings>=2.11.0",
]
